lower: move type conversion logic into `old.rs`
