+++ compiler/hash-lower/src/ty/new.rs
@@ -1,0 +1,2 @@
+//! Contains all of the logic that is used by the lowering process
+//! to convert types and [Ty]s into [IrTy]s.
